backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
httpx==0.27.2
httptools==0.7.1
idna==3.11
joblib==1.5.3
lightgbm==4.6.0
numpy==2.4.2
pandas==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
scikit-learn==1.6.1
scipy==1.17.1
shap==0.48.0
six==1.17.0
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0

# RAG dependencies (merged from Ai-Credit-Scoring)
nest-asyncio
llama-index-core>=0.14.0
llama-index-llms-openai
llama-index-llms-ollama
llama-index-embeddings-openai
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
llama-index-vector-stores-chroma
openai
tiktoken
ollama==0.6.1
chromadb
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
sqlalchemy>=2.0,<3
llama-index-llms-google-genai
